# api-documentation.yaml
openapi: 3.0.3
info:
  title: Organization Service API
  description: |
    API documentation for the Organization Service which manages organizations, patients, and users.
    
    ## Authentication
    All endpoints (except /health) require Bearer token authentication.
    Header: `Authorization: Bearer <your-token>`
    
    ## For SUPER_ADMIN
    Add `X-Organization-ID` header to specify target organization.
    
    ## For ORG_ADMIN
    Uses organization from token automatically. Do NOT send X-Organization-ID header.
  version: 1.0.0
  contact:
    name: WailSalutem Health Care
    email: support@wailsalutem.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.wailsalutem.com
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationID:
      name: X-Organization-ID
      in: header
      required: false
      description: Organization ID (required for SUPER_ADMIN, forbidden for ORG_ADMIN)
      schema:
        type: string
        format: uuid
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        keycloakUserId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
        orgId:
          type: string
          format: uuid
        orgSchemaName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - firstName
        - lastName
        - role
      properties:
        username:
          type: string
          example: caregiver.smith
        email:
          type: string
          format: email
          example: smith@healthcare.com
        firstName:
          type: string
          example: Sarah
        lastName:
          type: string
          example: Smith
        phoneNumber:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum: [CAREGIVER, MUNICIPALITY, INSURER]
          example: CAREGIVER
        temporaryPassword:
          type: string
          example: TempPass123!
        sendResetEmail:
          type: boolean
          default: false
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
    ResetPasswordRequest:
      type: object
      properties:
        temporaryPassword:
          type: string
          example: NewTempPass123!
        sendEmail:
          type: boolean
          example: true
    MessageResponse:
      type: object
      properties:
        message:
          type: string
paths:
...
  /organization/users:
    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user (CAREGIVER, MUNICIPALITY, INSURER)
      operationId: createUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    get:
      tags:
        - Users
      summary: List users
      description: List all users in organization
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /organization/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Get user details
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    patch:
      tags:
        - Users
      summary: Update user
      description: Update user information
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - Users
      summary: Delete user
      description: Soft delete user
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      responses:
        '204':
          description: User deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /organization/users/me:
    patch:
      tags:
        - Users
      summary: Update own profile
      description: Update authenticated user's own profile
      operationId: updateMyProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /organization/users/{id}/reset-password:
    post:
      tags:
        - Users
      summary: Reset user password
      description: Reset password for a user (requires X-Organization-ID for SUPER_ADMIN)
      operationId: resetPassword
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found